Use a state machine to manage ApplicationDraft states
Previously, the state of an ApplicationDraft was controlled by setting
`*_by` attributes. While this works as long as there is a small number
of states, this gets complicated quickly if the possible states
increases, as well as managing transitions between states.

When an ApplicationDraft transitions from the 'draft' state to the
'applied' state by calling `submit_application`, it sets the
`applied_at` column to either a supplied DateTime, or the current
DateTime.