WIP: new method for getting the points of a rateable
In general three kinds of objects might be rated:
- a team
- a student (actually a user)
- an application

All these combined give an average total number of points for the entire application (meaning: students, team and application combined).

This is a WIP because once again tests are missing and it is not used throughout the UI. It might be a good starting point for a discussion though.