ratings can now be saved again
Tried to stay as railsy as possible by setting up an association form Application as well.

The main difference model wise is that the data for each rating is now retrived from the serializeable hash and the hash is re-built before saveing. (Otherwise rails won't recognize what it needs to save. Maybe there is a better way to do this by marking data as changed, but this seems to work fine for now.)