added spec for TeamPerformance found 2 more edge cases
The edge cases: Buffer days at the beginning and end of the semester would only be taken into account for teams without any activites or comments. So if a team got a comment or activity before the sart of the semester it might be listed as worse off than if it got nothing.