Fixed export functionality (#626)
* Fixed export functionality

The dynamically created methods overwrote the param injected
into the method. This lead to the problem that the instance
could only export one thing. Every export after it was damaged
because of a wrong value in the season variable.

The navigation was kinda broken on all browsers because of
a wrong nesting of rows and widths of columns inside of the
row.

I replaced it completely with a new, cleaned up navigation and
made it look like the old one.

* fixed text shadow and border in navbar