Update task to flag applications
Compute values for "remote_team" flag based on the coordinates of the
students' locations rather then the city names. If the coordinates are
missing, the team is not marked as remote, since this is probably not
their fault.
The distances are calucated via the Haversine Formula (best practice for
this).