Align behavior of revoking mentor-favs
This aligns the behavor of revoking mentor-favs with revoking sign-offs:
setting the value back to "nil" instead of "false". I don't see a
benefit in having "false" and I think it just makes things easier to
handle if we instead of setting it to "nil" simply remove it completely
from the application_data hash (will result in "nil" anyways and also
works if it's already "nil").