remove global css, use boostrap markup instead
Prior to this commit a few global custom margin directives for the btn class and related classes existed that messed up spacing in various forms.
As far as I can tell these where originally introduced to deal with local padding issues in the search form and user show form.
The first issue was fixed by using standard boostrap form componets for the search form. The latter issue did not seem to be an obvious problem anymore anyway.