Fix deprecation warnings for old AR::Dirty methods :rocket:
Parts of AR::Dirty was deprecated, but there's still a bug when these
warnings trigger. In our case it was all over the place plus the warning
IS WRONG: it tells to switch to a method that does not fit the situation
and does something completely different.

Here is the commit on Rails: https://github.com/rails/rails/commit/16ae3db5a5c6a08383b974ae6c96faac5b4a3c81

See DHH's comment where he also mentions the current bug with
deprecations warnings and encourages users to instead just use the new
API. So I switched to the new API for use.