Change import format to tabs separated :sparkles:
Tabs separated is a more robust solution for csv, especially when there are comma's within the fields (like we have in date fields: Aug 12, 2017). Normally this is solved by setting the column separator in .csv to ';', but that seems to be depending on the system (windows/mac), the geographical region and the spreadsheet app (google, numbers, excell). That is too many choices :joy: