Move mailing stuff into orgs namespace (#793)
* Controller specs :white_check_mark:

* Copy, move and adjust files into orga namespace :hammer:

* Add mailings to orga dashboard :nail_care:

* Remove admin abilities :fire:

Access is guarded by the orga base controller

* Finetune cooperation between the 2 controllers :art:

- Simplify redirects
- Make routing so that it gives meaningful notices on authorisation
- Use partials for shared views

* Add breadcrumbs :nail_care:

* Clean up specs, add some examples :white_check_mark:

* Move submission stuff into orga namespace

* Typo's and whitespaces :pencil:

* Remove a lot of controller `before_actions` :fire: :ok_hand:

* Shared view partials, change ivar to var :ok_hand:

* :construction: New specs, still with memo comments

* Add ivar declaration to index action directly

* Restore mailing partial in orga namespace :bug:

and use var in table instead of ivar

* Restore Cancancan's redirection :art:

I made up a complicated solution with redirecting via the orga base controller, but cancancan already provides that solution
 with the load_and_authorize_resouce method.
 For the actions where we want to show a neat notice instead of exceptions, we need to add a declaration to the controller, and we need all the `resources: mailings` routes without restrictions.

* Remove redundant :mailing_params in public controller :fire:

* Share _mailing partial between namespaces :art:

- Share partial
- Make subject link depending on authorisation

* Adjust path for show view :bug:

* Add validation for 'to:' for dev happiness :gift_heart:

Without the 'to:' set, sending the email silently fails.

* Solved mystery with failing shared_context

Restored 'build_stubbed user' instead of 'build user' in the shared context file :bug:

* Cleaning up specs :art:

* Solving bug in shared_context, yes, really! :bug:

* Remove redundant user :fire:

* Remove redundant authorize calls :fire:

For the orga/mailings, authorisation is done via the base controller
For the public mailings, the general authorisation doesn't work on the :index action. Only there the explicit call is needed.

* Add validation for subject :art:

Without subject, we don't have access to the email without looking up the id. With subject, we have a clickable link in the table.

* Remove redundant else clause in mailing params; set default for 'from' :bug:

The else clause is never fired. Therefore, the default from address is not set either.
Now, the default is set in the form and the params method simplified.

* Afterburners :lipstick:

* Revert changing the symbol :bug:

* Finetuning loading and authorizing of resource :zap: