Update contribution guide with the ping method (#815)
* Add "the ping method" to the contribution guide

As agreed upon on Slack, give the ping method a try for PRs.

* Remove the "needs review" label from the guide

This removes the "needs-review" label from the guide (I would afterwards
of course also remove it from the Repo itself...). Since up to now we
are not using it anymore and Github's new feature of requesting PR
reviews basically makes it obsolete, I think it's also a bit obsolete in
the guide - just makes things more complex there.