Migrate old team "kinds" to new datamodel (#946)
* Fix specs for team associations

Also: a bit of reordering and removing of the no-longer working
"conferences" association.

* Update team accepted definition

This now includes the legacy scopes "voluntary" and "sponsored".

* Remove duplicated team "selected" scope

This was the same as "accepted"... So no need for keeping it.

* Update visible scope

This now also includes legacy teams (even though it may no longer be
relevant...)

* Use team's accepted scope in other queries

This is supposed to support and include the legacy "kinds". Also: moved
a method from a random position in the user model into the class block.

* Remove commented code

* Simplify team.by_season

The scope was too overkill for what it was actually doing. I removed the
things that weren't used to slim down the team model.