What a difference a bang makes. Thanks @pgaspar
Now that devise's confirmation delivered_later is moved to the user model, users created in specs (not going through OAuth) will trigger deliver_later as well. That causes the miscount in this spec; Before, the users confirmation mail were delivered_now, and didn't count as enqueued jobs.
Banging the let draft makes that the user confirmations are sent outside of the example, and thus cleared before the example runs.