Update team and project controller
Update the controllers to use the teams' project_id instead of the old
project_name column.
For the orga teams controllers, I also updated the form to enable choosing
the project from a dropdown list. It's questionable though whether this
should be possible for THE TEAMS themselves, so I removed it from this
form.

closes #892