[#892] Migrate `team.project_name` to `team.project_id` (#1022)
* Require applications to have an application_draft

This makes the association between Application and ApplicationDraft
required. While not super necessary, this is maybe helpful to ensure the
workflow of moving from an application-draft to an actual application.

* Update presence validations for the application model

With the association to teams being required, having a presence
validation just duplicates the error message on the model.

* Remove project_name column from teams table

This also migrates all the existing data "over" to use the project_id
instead.
Plus: I also removed the rake task, since it apparently wasn't run on
production anyways.

* Update team model to use project_id

Update the team model to make use of the project_id column, rather than
the old project_name one.

* Adapt project model to use the teams' project_ids

Instead of the project_name column, also query based on the project_id
here.

* Update team and project controller

Update the controllers to use the teams' project_id instead of the old
project_name column.
For the orga teams controllers, I also updated the form to enable choosing
the project from a dropdown list. It's questionable though whether this
should be possible for THE TEAMS themselves, so I removed it from this
form.

closes #892

* Change window size for headless chrome driver

* Fix typo in specs

Remove unnecessary assignment.

* Enable new capybara driver to do screenshots

The custom driver does not know how to do this by default.

* Remove unnecessary team model spec

There's a presence validation on the name attribute, so testing for
cases where a team does not have a name is obsolete.

* Display the project on the team edit form

We just display a readonly field.

* Remove project_id left-overs in specs

* Update team form project field

Disable the field instead of making it readonly (otherwise the field
would still be tabable and the value submitted to the server - where we
ignore it).