Remove some logic from the seeds file
Seeding the database would somehow hang forever for me with the logic
for "resetting" the delivery method again. When removing it like this,
it works like charm. There's no sideeffekt, since the actual app gets
boot up in a separate process anyways, and there we have the setting
hardcoded again.