Delete Comments with User (#1028)
* Make comment associations required

We in probably all cases want to ensure a comment has a user and a
commentable.

* Remove comments when user deletes their accout

This makes all the comments be deleted with the author of the comment
deleting their account.

* Migrate existing comments

We delete the comments where the user no longer exists in the database.

* Adapt the view helpers for comment authors

The helper no longer has to care about deleted users.

* Remove some logic from the seeds file

Seeding the database would somehow hang forever for me with the logic
for "resetting" the delivery method again. When removing it like this,
it works like charm. There's no sideeffekt, since the actual app gets
boot up in a separate process anyways, and there we have the setting
hardcoded again.

* Remove wip tag :see_no_evil: