Update Rubocop strategy (#1066)
* Remove old Rubocop config

Remove basically everything from the old Rubocop configuration. We're
just leaving some option to display the output including information
about the documentation for certain cops and to enable cops related to
Rails.

* Add rubocop-rspec

The tests are part of the code we're maintaining and equally important
as the business logic - so we want to care about them.

* Generate a todo file for Rubocop

* Have Rubocop fix some small alignment things

This was too small to be included in the todo list. Therefore quickly
ran Rubocop with auto correct.