Rubocop fixes pt. 1 - the 🚀 edition (#1088)
* Apply cop to order Gemfile

This applies the very basic cop to check the alphabetic ordering -
within groups - of the Gemfile. I reordered some things, so that it
makes sense logically.

* Apply cop for dynamic attributes in factories

This ensures one uses the block syntax for dynamic attributes in
factories (otherwise the result will always be the same).

* Apply cop to add a blank line after the magic comment

This ensures a blank line after the magic frozen string comment.

* Apply cop to ensure blank lines between method definitions

* Apply cop to remove unnecessary blank lines

This is simply removing extra blank lines that slipped in over the
years.

* Apply block to remove blank lines at block-start

This removes blank lines around the body of a block. Especially useful
when writing tests, to easier find out where the "it" blocks are.

* Apply cop to remove empty class body start

These cops ensure a class body does not start with a blank line, a
module body does not start with a blank line, etc.

* Apply cop to ensure spaces before blocks

This cop makes sure blocks have a space before their "{}" (same for
lambdas of course).

* Apply cop to remove spaces after array brackets

* Run cop to add missing spaces in hashes

* Run cop to clean up extra space in brackets

* Run cop to clean up extra space in percentage literals