Fix flaky feature spec
There was a feature spec creating another season... The season model is
a bit of a pain due to its logic around the `current` season and tending
to run into name conflicts. Therefore, the test now ensures we're really
dealing with a "past" season and won't run into naming conflicts with
the factory.