Rubocop fixes 2 (#1089)
* Remove duplicated methods

With the help of Rubocop, detecting some duplicated methods and removing
the first (in Ruby the second definition overrides the first - so the
second was the one to be called).

* Fix class methods marked as private

There were some class methods marked as "private" but not actually
private.

On the side: also remove a duplicated and therefore useless "private"
modifier key.

* Fix flaky feature spec

There was a feature spec creating another season... The season model is
a bit of a pain due to its logic around the `current` season and tending
to run into name conflicts. Therefore, the test now ensures we're really
dealing with a "past" season and won't run into naming conflicts with
the factory.